requirements.txt
asgiref==3.9.1
Django==5.2.6
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.17
djangorestframework==3.16.1
dotenv==0.9.9
pillow==12.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
